A Data-Driven Analysis for market Expansion
GlobalBite is a data-driven analysis project that aims to identify potential new markets for the company. The project is part of a broader market research campaign aimed at expanding the company's presence in major cities worldwide. GlobalBite focuses on the cost of living, expense, and salary data to pinpoint potential cities for deeper research into potential new markets for the company.